pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "4d-paper"
version = "1.0.0"
authors = [
  { name="4D-Paper Team", email="team@4dpaper.example.com" }
]
description = "Dynamic 4D Academic Paper System with multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.1",
  "uvicorn>=0.24.0",
  "pydantic>=2.4.2",
  "pandas>=2.1.3",
  "numpy>=1.26.0",
  "matplotlib>=3.8.1",
  "h5py>=3.10.0",
  "cryptography>=41.0.7",
  "python-jose>=3.3.0",
  "markdown2>=2.4.10",
  "reportlab>=3.6.12",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "black>=23.0",
  "isort>=5.0",
  "flake8>=6.0",
]